lines: lines mask, tests, and properties
* remove nan_depth_value property and setter
* create appropriate lambda function to replace replace_depth
* place lines mask in lines class as a class function
* remove checking for ._data type (it is hidden and is set by code so we know it is dataframe)
* separate start time and end time checking
* revise the usage of lines_mask in tests so that we have just lines.mask(...)
