+++ echoregions/tests/test_line.py
@@ -66,9 +66,8 @@ def test_replace_nan_depth():
     """
     Test replacing NaN values in line.
     """
-    lines = er.read_evl(evl_path)
+    lines = er.read_evl(evl_path, nan_depth_value=20)
     lines.data.loc[0, "depth"] = -10000.99  # Replace a value with the one used for nans
-    lines.nan_depth_value = 20
     lines.replace_nan_depth(inplace=True)
     assert lines.data.loc[0, "depth"] == 20
 
@@ -81,7 +80,7 @@ def test_lines_mask():
 
     da_Sv = xr.open_dataset(os.path.join(data_dir, "x1_test.nc")).Sv
 
-    M = er.lines_mask(da_Sv, lines)
+    M = lines.mask(da_Sv)
     M.plot(yincrease=False)
     # from matplotlib import pyplot as plt
     # plt.show()
